KAN-69: Rename: join response dto
